refactor(4b): move runtime types (PaneRef et al) to @cockpit/shared
