+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 // @cockpit/shared — pure types + leaf utilities + config.
 export * from "./config.js";
+export * from "./types/runtime.js";
 export * from "./types/control.js";
 export * from "./types/projection.js";
 export * from "./types/workspaces.js";
